requirements.txt
alembic==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
ndjson==0.3.1
numpy==1.19.4
pandas==1.1.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.22
